requirements.txt
streamlit==1.28.1
streamlit-option-menu==0.3.6
qrcode[pil]==7.4.2
pillow==10.1.0
pyzbar==0.1.9
opencv-python-headless==4.8.1.78
numpy==1.24.3
pandas==2.0.3
gspread==5.11.3
oauth2client==4.1.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
opencv-python-headless==4.8.1.78
numpy==1.26.2